Add an "entry hole" for the block code that will be executed
This commit greatly refactors the logic for block positioning.
Previously, when a hole contained a block, the hole was invisible
and the block's SVG was a child of the parent block's SVG. Now, the
block's SVG is a child of the hole's SVG. As a result, the block's
position is (0, 0) when it is a child and (x, y) when it is a free
script in the workspace area. So, when converting between local
and global coordinates, one must use holes instead of blocks.