Change the API for specifying custom widgets.
I make a distinction between the widgets that appear in block templates from
which one can drag blocks and the widgets that appear in the actual blocks.
Previously, I had made the text input that appeared in the template actually
a regular text widget, but this looked ugly. This commit makes the text an
actual input; editing the input in the template changes the value of the input
of the block that gets dragged out. In the process of making this change, I
replaced the `text_input` function in Bexp.Syntax with the more general
`widget` function, a thin wrapper over the Widget data constructor.